Replace use of String[] array with Person class
AddressBook uses a String[] array with 3 elements to store the name,
phone and email of a single person respectively. Such usage of an array
is undesirable as it constrains the data fields of a person to always be
Strings, and it is also not type-safe as there is no way to
differentiate such "person arrays" from any other 3-element string
arrays. Furthermore, it makes the code hard to read as it is not readily
apparent that a string array is actually a "person array".

Refactor the code to use a Person class instead. This makes it explicit
that we are dealing with Person objects, allows members of other types
to be added, and improves the readability of the code.

Since the Person class is an internal implementation detail of the
AddressBook, it is implemented as a private nested class.